Create OIMS Backend update script
this script runs in ArcGIS Online and reads in the various feature services and csv files and performs operations on them. It may be possible to run it as a stand alone script but I have not had a chance to look at that yet. this script requires testing for bugs